+++ CHANGELOG.md
@@ -6,6 +6,18 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [Checkpoint] - 2025-12-04 12:56:49
+
+### Changed Files
+- `src/features/comps/CompFilters.tsx`
+- `src/features/hierarchy/components/AgentTable.tsx`
+- `src/features/recruiting/components/FilterDialog.tsx`
+
+### Statistics
+```
+ 3 files changed, 101 insertions(+), 34 deletions(-)
+```
+
 ## [Checkpoint] - 2025-12-04 11:31:45
 
 ### Changed Files